move realignment plaquettes definition into spec genrators
