+++ src/tqec/compile/specs/library/generators/fixed_bulk_test.py
@@ -1,8 +1,10 @@
 import stim
 
-from tqec.plaquette.library.hadamard import make_fixed_bulk_realignment_plaquette
+from tqec.compile.specs.library.generators.fixed_bulk import (
+    make_fixed_bulk_realignment_plaquette,
+)
 from tqec.plaquette.qubit import SquarePlaquetteQubits
-from tqec.plaquette.rpng.rpng import XYZBasis
+from tqec.plaquette.rpng.rpng import PauliBasis
 from tqec.utils.enums import Basis, Orientation
 
 
@@ -12,7 +14,7 @@ def test_fixed_bulk_realignment_plaquette() -> None:
         z_orientation=Orientation.VERTICAL,
         mq_reset=Basis.X,
         mq_measurement=Basis.Z,
-        debug_basis=XYZBasis.X,
+        debug_basis=PauliBasis.X,
     )
     assert plaquette.qubits == SquarePlaquetteQubits()
     assert len(plaquette.circuit.schedule) == 6
@@ -31,14 +33,14 @@ TICK
 MZ 4
 H 0 1 2 3
 """)
-    assert plaquette.debug_information.get_basis() == XYZBasis.X
+    assert plaquette.debug_information.get_basis() == PauliBasis.X
 
     plaquette = make_fixed_bulk_realignment_plaquette(
         stabilizer_basis=Basis.Z,
         z_orientation=Orientation.HORIZONTAL,
         mq_reset=Basis.Z,
         mq_measurement=Basis.X,
-        debug_basis=XYZBasis.Z,
+        debug_basis=PauliBasis.Z,
     )
     assert plaquette.qubits == SquarePlaquetteQubits()
     assert len(plaquette.circuit.schedule) == 6
@@ -57,4 +59,4 @@ TICK
 MX 4
 H 0 1 2 3
 """)
-    assert plaquette.debug_information.get_basis() == XYZBasis.Z
+    assert plaquette.debug_information.get_basis() == PauliBasis.Z
